Add support for Mac OS X 10.4: - Replace uses of NSInteger with int. - Remove unnecessary call to sqlite3_clear_bindings().
To build on 10.4, a newer version of SQLite is required. No published API was modified.



git-svn-id: https://pldatabase.googlecode.com/svn/trunk@238 3603cd71-484c-0410-acc4-09e5374cfb26

+++ Classes/PLDatabaseMigrationDelegate.h
@@ -53,6 +53,6 @@
  * why the migration could not be completed. If no error occurs, this parameter will be left unmodified. You may specify nil for this
  * parameter, and no error information will be provided.
  */
-- (BOOL) migrateDatabase: (id<PLDatabase>) database currentVersion: (NSInteger) currentVersion newVersion: (NSInteger *) newVersion error: (NSError **) outError;
+- (BOOL) migrateDatabase: (id<PLDatabase>) database currentVersion: (int) currentVersion newVersion: (int *) newVersion error: (NSError **) outError;
 
 @end
